代理端接口

+++ yuepei-system/src/main/java/com/yuepei/system/domain/UserLeaseOrder.java
@@ -66,13 +66,13 @@ public class UserLeaseOrder extends BaseEntity {
     private String restoreAddress;
 
     /** 应付金额 */
-    private Long price;
+    private BigDecimal price;
 
     /** 支付方式 */
     private String payType;
 
     /** 实付金额 */
-    private Long netAmount;
+    private BigDecimal netAmount;
 
     /** 优惠券额度 */
     private Long couponPrice;
